+++ packages/wasm/lib/image-xobjects.js
@@ -1,0 +1,5 @@
+/**
+ * Image XObjects created by Wasm writers and modifiers. Resource mapping
+ * accepts these in place of an object ID, as native does.
+ */
+export var imageXObjects = new WeakSet();